Do not allow the removal of past option planning periods

+++ module/Activity/src/Controller/AdminOptionController.php
@@ -110,6 +110,13 @@ class AdminOptionController extends AbstractActionController
             $optionCreationPeriod = $this->activityCalendarService->getOptionCreationPeriod($optionCreationPeriodId);
 
             if (null !== $optionCreationPeriod) {
+                if ($optionCreationPeriod->getEndPlanningTime() < new DateTime('now')) {
+                    return $this->redirectWithMessage(
+                        false,
+                        $this->translator->translate('Past option planning periods cannot be removed.'),
+                    );
+                }
+
                 $this->activityCalendarService->deleteOptionCreationPeriod($optionCreationPeriod);
 
                 return $this->redirectWithMessage(
